BUGFIX: Make selects in `stack config host interface` case insensitive when looking up MACs.
If a MAC address has uppercase letters in the database and if `stack config host interface`
is called with MAC addresses with lowercase letters, the code wouldn't find a match,
when, in fact, the MACs did match.

This code is called during kickstart file generation (by `profile.cgi`) where it adds new
MAC addresses to the database if those MAC addresses didn't previously exist. Without this
change, the existing MAC address for the `private` interface wouldn't match, which would
result in no new MAC addresses being added to the database.