FEATURE: Improved cluster data output in `report system`
The `stack report system` command now outputs a section in the end of the output containing all the `list` data collected during the commands. I also added data for `list cart` because it seemed to be missing.

To use this in other tests, simply include the new `report_output` fixture and call it inside of your test function, passing it a "title" for the output and the output itself, both as strings.